change track timescale to unsigned int, fix warning
Originally committed as revision 20651 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavformat/movenc.c
@@ -58,7 +58,7 @@ typedef struct MOVIentry {
 typedef struct MOVIndex {
     int         mode;
     int         entry;
-    long        timescale;
+    unsigned    timescale;
     uint64_t    time;
     int64_t     trackDuration;
     long        sampleCount;
